refactor: range le répertoire infra (000)
