+++ src/server/infrastructure/accès_données/indicateur/IndicateurRandomRepository.ts
