+++ src/server/infrastructure/accès_données/ministère/MinistèreInMemoryRepository.ts
