+++ src/server/infrastructure/accès_données/synthèseDesRésultats/SynthèseDesRésultatsRandomRepository.ts
