+++ src/server/infrastructure/accès_données/synthèseDesRésultats/SynthèseDesRésultatsSQLRepository.ts
@@ -1,6 +1,6 @@
 import { synthese_des_resultats, PrismaClient } from '@prisma/client';
 import SynthèseDesRésultatsRepository from '@/server/domain/synthèseDesRésultats/SynthèseDesRésultatsRepository.interface';
-import { CODES_MAILLES } from '@/server/infrastructure/maille/mailleSQLParser';
+import { CODES_MAILLES } from '@/server/infrastructure/accès_données/maille/mailleSQLParser';
 import { Maille } from '@/server/domain/maille/Maille.interface';
 import { CodeInsee } from '@/server/domain/territoire/Territoire.interface';
 import { DétailsCommentaire } from '@/server/domain/commentaire/Commentaire.interface';
